Rename package to clickhouse
