+++ src/main/resources/application.properties.EXAMPLE
@@ -9,10 +9,10 @@ spring.datasource.mysql.url=jdbc:mysql://localhost:3306/cbioportal?useSSL=false
 spring.datasource.mysql.username=cbio
 spring.datasource.mysql.password=P@ssword1
 spring.datasource.mysql.driver-class-name=com.mysql.jdbc.Driver
-spring.datasource.columnar.url=jdbc:ch://localhost:8443/cbioportal
-spring.datasource.columnar.username=dummy
-spring.datasource.columnar.password=dummy
-spring.datasource.columnar.driver-class-name=com.clickhouse.jdbc.ClickHouseDriver
+spring.datasource.clickhouse.url=jdbc:ch://localhost:8443/cbioportal
+spring.datasource.clickhouse.username=dummy
+spring.datasource.clickhouse.password=dummy
+spring.datasource.clickhouse.driver-class-name=com.clickhouse.jdbc.ClickHouseDriver
 spring.jpa.database-platform=org.hibernate.dialect.MySQL5InnoDBDialect
 
 # this should normally be set to false. In some cases you could set this to true (e.g. for testing a feature of a newer release that is not related to the schema change in expected db version above):
