+++ src/main/resources/org/cbioportal/persistence/mybatisclickhouse/StudyViewMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="org.cbioportal.persistence.mybatiscolumnar.StudyViewMapper">
+<mapper namespace="org.cbioportal.persistence.mybatisclickhouse.StudyViewMapper">
 
     <!-- for /filtered-sample/fetch (returns Sample objects) -->
     <select id="getFilteredSamples" resultType="org.cbioportal.model.Sample">
