+++ src/test/java/org/cbioportal/persistence/mybatisclickhouse/config/MyBatisConfig.java
@@ -1,4 +1,4 @@
-package org.cbioportal.persistence.mybatiscolumnar.config;
+package org.cbioportal.persistence.mybatisclickhouse.config;
 
 import org.mybatis.spring.SqlSessionFactoryBean;
 import org.mybatis.spring.annotation.MapperScan;
@@ -12,18 +12,18 @@ import javax.sql.DataSource;
 import java.io.IOException;
 
 @TestConfiguration
-@MapperScan("org.cbioportal.persistence.mybatiscolumnar")
+@MapperScan("org.cbioportal.persistence.mybatisclickhouse")
 public class MyBatisConfig {
 
     @Bean
-    @ConfigurationProperties("spring.datasource.columnar")
-    public DataSourceProperties columnarDatSourceProperties() {
+    @ConfigurationProperties("spring.datasource.clickhouse")
+    public DataSourceProperties clickhouseDatSourceProperties() {
         return new DataSourceProperties();
     }
 
     @Bean
     public DataSource dataSource() {
-        return columnarDatSourceProperties()
+        return clickhouseDatSourceProperties()
             .initializeDataSourceBuilder()
             .build();
     }
@@ -32,8 +32,8 @@ public class MyBatisConfig {
     public SqlSessionFactoryBean sqlColumnarSessionFactory(ResourceLoader resourceLoader, DataSource dataSource) throws IOException {
         SqlSessionFactoryBean sessionFactory = new SqlSessionFactoryBean();
         sessionFactory.setDataSource(dataSource);
-        var studyViewMapperResource = resourceLoader.getResource("classpath:org/cbioportal/persistence/mybatiscolumnar/StudyViewMapper.xml") ;
-        var studyViewFilterMapperResource = resourceLoader.getResource("classpath:org/cbioportal/persistence/mybatiscolumnar/StudyViewFilterMapper.xml");
+        var studyViewMapperResource = resourceLoader.getResource("classpath:org/cbioportal/persistence/mybatisclickhouse/StudyViewMapper.xml") ;
+        var studyViewFilterMapperResource = resourceLoader.getResource("classpath:org/cbioportal/persistence/mybatisclickhouse/StudyViewFilterMapper.xml");
         sessionFactory.setMapperLocations(
             studyViewMapperResource,studyViewFilterMapperResource
         );
